Move MessagesBuffer to separate file, add assert, test error during flush, and account for new log message in max size test
